+++ src/NeatMapper/AssemblyInfo.cs
@@ -1,5 +1,7 @@
-﻿using System.Runtime.CompilerServices;
+﻿using System;
+using System.Runtime.CompilerServices;
 
+[assembly: CLSCompliant(true)]
 [assembly: InternalsVisibleTo("NeatMapper.EntityFrameworkCore")]
 [assembly: InternalsVisibleTo("NeatMapper.EntityFrameworkCore.Tests")]
 [assembly: InternalsVisibleTo("NeatMapper.Transitive")]